pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer is a website that analyzes the specified pages for SEO suitability by analogy"
authors = ["Faizik12 <zerg9119@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]
repository = "https://github.com/Faizik12/python-project-83"


[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
validators = "^0.22.0"
types-psycopg2 = "^2.9.21.13"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
mypy = "^1.5.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
